% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{run_model}
\alias{run_model}
\title{Run Model}
\usage{
run_model(params, control, param.file)
}
\arguments{
\item{params}{List of model parameters.}

\item{control}{List of model controls.}
}
\value{
A 3-column matrix containing the predator and prey populations 
  for each timestep.
}
\description{
Run the Lokta-Volterra predator-prey model.
}
