% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestutils.r
\name{write_instructions}
\alias{write_instructions}
\title{Write Instructions File}
\usage{
write_instructions(ins.file, n)
}
\arguments{
\item{ins.file}{Path to the instructions file. If the file already exists,
it will be overwritten.}

\item{n}{Number of observation rows to write.}
}
\description{
Write Instructions File
}
