% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{write_outputs}
\alias{write_outputs}
\title{Write Model Outputs}
\usage{
write_outputs(d, out.file)
}
\arguments{
\item{d}{The model output data.}

\item{out.file}{Path to the output file. If the file already exists,
it will be overwritten.}
}
\description{
Write the model outputs to a file.
}
